change return type of ConvertGBtoB from float64 to uint64

+++ utils.go
@@ -28,6 +28,6 @@ func ConvertMBToB(size uint64) uint64 {
 	return size * uint64(MB)
 }
 
-func ConvertGBToB(size float64) float64 {
-	return size * GB
+func ConvertGBToB(size float64) uint64 {
+	return uint64(size * GB)
 }
